Add tests and models for UVL and FAMA
