+++ models/fama/WeaFQAs.fama
@@ -1,0 +1,65 @@
+%Relationships
+DatabaseType: [1,1]{Relational NoSQL};
+EncryptionImplementations: [1,1]{JavaxCrypto BouncyCastle};
+Encryption: EncryptionAlgorithms [KeySize] EncryptionImplementations EncryptionOperations Mode Padding;
+Security: [1,7]{Encryption Authentication Hashing Confidentiality Privacy Integrity Non_Repudiation};
+PasswordType: [1,1]{Numeric Alphanumeric SpecialChars};
+ContextualHelpImplementations: JavaWizardAPI;
+ErrorLog: [TraceFile];
+Persistence: [1,2]{FileStorage DatabaseStorage};
+EncryptionAlgorithms: [1,1]{Asymmetric Symmetric};
+Hashing: HashingImplementations HashingOperations HashingAlgorithms;
+ExecutionControl_1: [1,4]{Undo Redo Cancel GoBack};
+Level: [1,5]{Trace Debug Info Warning Error};
+Performance: Caching;
+UserType: [1,4]{Beginner Intermediate Advanced Expert};
+CachingOperations: [1,1]{CacheStore CacheQuery CacheUpdate};
+DeviceAware: [1,2]{BatteryStatus IRSensor};
+DataType: [1,4]{Integers Strings Floats Objects};
+CachingImplementations: [1,1]{JavaCachingSystem Caffeine};
+AuthenticationMechanism: [1,1]{UserPassword DigitalCertificate Biometric SocialID Pin MatrixCode};
+DatabaseStorageOperations: [1,1]{Store Query};
+LocationAware: [1,2]{GPS WiFiLocation};
+AuthenticationOperations: Authenticate;
+Authentication: AuthenticationImplementations AuthenticationOperations AuthenticationMechanism [PasswordLength] [PasswordType] [LimitedSession] [AuthenticationLogging] [AnonymousUser];
+FileStorage: FileStorageOperations FileType FileStorageImplementations [TemporalFile];
+FQAs: [1,5]{Security Usability Persistence ContextAwareness Performance};
+Symmetric: [1,1]{AES DESede Blowfish DES};
+EncryptionOperations: [1,1]{Encrypt Decrypt};
+DigitalCertificate: [PseudonymousCertificate];
+Asymmetric: [1,1]{RSA DSA ECDSA};
+FileStorageImplementations: JavaIO;
+HashingImplementations: [1,1]{JavaSecurityMessageDigest Guava};
+Logging: LoggingOperations LoggingImplementations Handler Format Level [CypheredMessages] [HistoryLog] [ErrorLog];
+DatabaseStorageImplementations: [1,1]{MySQL MongoDB OracleDB};
+Handler: [1,1]{ConsoleLog FileLog DatabaseLog};
+FileStorageOperations: [1,1]{CreateFile ModifyFile DeleteFile};
+HashingAlgorithms: [1,1]{MD5 SHA1 SHA256 SHA512};
+DatabaseStorage: DatabaseStorageOperations DatabaseStorageImplementations DatabaseType DataType [SecureDatabase];
+ContextualHelpOperations: ShowHelp;
+ContextAwareness: [1,5]{LocationAware TimeAware DeviceAware UserAware EnvironmentAware};
+CacheType: [1,1]{LocalCache RemoteCache};
+LoggingOperations: Log;
+Format: [1,1]{TXT XML HTML JSON};
+Feedback: [1,4]{InteractionFeedback ProgressFeedback SystemStatusDisplay Alerts};
+Padding: [1,1]{PKCS5 PKCS1 OAEP};
+HashingOperations: Hash;
+ExecutionControl: ExecutionControl_1 [Shortcuts];
+AuthenticationImplementations: [1,1]{JAAS SpringSecurity};
+LoggingImplementations: [1,1]{Log4J LogBack JavaUtilLogging SimpleImplementationSLF4J};
+Mode: [1,1]{CBC ECB CFB CTR CTS};
+ContextualHelp: ContextualHelpOperations ContextualHelpImplementations HelpType UserType [AuthenticatedUser];
+Caching: CachingOperations CachingImplementations MaxSizeCache CacheType [MaxLifeSeconds] [EternalObjects];
+FileType: [1,1]{Binary TXTFile XMLFile JSONFile};
+Internationalization: Locale;
+HelpType: [1,2]{Tutorial Wizard};
+Usability: [1,5]{Logging ContextualHelp ExecutionControl Internationalization Feedback};
+
+%Constraints
+BatteryStatus REQUIRES TimeAware;
+BatteryStatus REQUIRES Alerts;
+TraceFile REQUIRES FileStorage;
+SimpleImplementationSLF4J EXCLUDES XML;
+AuthenticatedUser REQUIRES Authentication;
+ErrorLog REQUIRES Error;
+HistoryLog REQUIRES Trace;
