+++ models/uvl/WeaFQAs.uvl
@@ -1,0 +1,262 @@
+namespace FQAs
+
+features
+	FQAs	
+		or
+			Security	
+				or
+					Encryption	
+						mandatory
+							EncryptionAlgorithms	
+								alternative
+									Asymmetric	
+										alternative
+											RSA
+											DSA
+											ECDSA
+									Symmetric	
+										alternative
+											AES
+											DESede
+											Blowfish
+											DES
+							EncryptionImplementations	
+								alternative
+									JavaxCrypto
+									BouncyCastle
+							EncryptionOperations	
+								alternative
+									Encrypt
+									Decrypt
+							Mode	
+								alternative
+									CBC
+									ECB
+									CFB
+									CTR
+									CTS
+							Padding	
+								alternative
+									PKCS5
+									PKCS1
+									OAEP
+
+						optional
+							KeySize
+					Authentication	
+						mandatory
+							AuthenticationImplementations	
+								alternative
+									JAAS
+									SpringSecurity
+							AuthenticationOperations	
+								mandatory
+									Authenticate
+							AuthenticationMechanism	
+								alternative
+									UserPassword
+									DigitalCertificate	
+										optional
+											PseudonymousCertificate
+									Biometric
+									SocialID
+									Pin
+									MatrixCode
+
+						optional
+							PasswordLength
+							PasswordType	
+								alternative
+									Numeric
+									Alphanumeric
+									SpecialChars
+							LimitedSession
+							AuthenticationLogging
+							AnonymousUser
+					Hashing	
+						mandatory
+							HashingImplementations	
+								alternative
+									JavaSecurityMessageDigest
+									Guava
+							HashingOperations	
+								mandatory
+									Hash
+							HashingAlgorithms	
+								alternative
+									MD5
+									SHA1
+									SHA256
+									SHA512
+					Confidentiality
+					Privacy
+					Integrity
+					Non_Repudiation
+			Usability	
+				or
+					Logging	
+						mandatory
+							LoggingOperations	
+								mandatory
+									Log
+							LoggingImplementations	
+								alternative
+									Log4J
+									LogBack
+									JavaUtilLogging
+									SimpleImplementationSLF4J
+							Handler	
+								alternative
+									ConsoleLog
+									FileLog
+									DatabaseLog
+							Format	
+								alternative
+									TXT
+									XML
+									HTML
+									JSON
+							Level	
+								or
+									Trace
+									Debug
+									Info
+									Warning
+									Error
+
+						optional
+							CypheredMessages
+							HistoryLog
+							ErrorLog	
+								optional
+									TraceFile
+					ContextualHelp	
+						mandatory
+							ContextualHelpOperations	
+								mandatory
+									ShowHelp
+							ContextualHelpImplementations	
+								mandatory
+									JavaWizardAPI
+							HelpType	
+								or
+									Tutorial
+									Wizard
+							UserType	
+								or
+									Beginner
+									Intermediate
+									Advanced
+									Expert
+
+						optional
+							AuthenticatedUser
+					ExecutionControl	
+						mandatory
+							ExecutionControl_1	
+								or
+									Undo
+									Redo
+									Cancel
+									GoBack
+
+						optional
+							Shortcuts
+					Internationalization	
+						mandatory
+							Locale
+					Feedback	
+						or
+							InteractionFeedback
+							ProgressFeedback
+							SystemStatusDisplay
+							Alerts
+			Persistence	
+				or
+					FileStorage	
+						mandatory
+							FileStorageOperations	
+								alternative
+									CreateFile
+									ModifyFile
+									DeleteFile
+							FileType	
+								alternative
+									Binary
+									TXTFile
+									XMLFile
+									JSONFile
+							FileStorageImplementations	
+								mandatory
+									JavaIO
+
+						optional
+							TemporalFile
+					DatabaseStorage	
+						mandatory
+							DatabaseStorageOperations	
+								alternative
+									Store
+									Query
+							DatabaseStorageImplementations	
+								alternative
+									MySQL
+									MongoDB
+									OracleDB
+							DatabaseType	
+								alternative
+									Relational
+									NoSQL
+							DataType	
+								or
+									Integers
+									Strings
+									Floats
+									Objects
+
+						optional
+							SecureDatabase
+			ContextAwareness	
+				or
+					LocationAware	
+						or
+							GPS
+							WiFiLocation
+					TimeAware
+					DeviceAware	
+						or
+							BatteryStatus
+							IRSensor
+					UserAware
+					EnvironmentAware
+			Performance	
+				mandatory
+					Caching	
+						mandatory
+							CachingOperations	
+								alternative
+									CacheStore
+									CacheQuery
+									CacheUpdate
+							CachingImplementations	
+								alternative
+									JavaCachingSystem
+									Caffeine
+							MaxSizeCache
+							CacheType	
+								alternative
+									LocalCache
+									RemoteCache
+
+						optional
+							MaxLifeSeconds
+							EternalObjects
+
+constraints
+	TimeAware | !BatteryStatus
+	Alerts | !BatteryStatus
+	FileStorage | !TraceFile
+	!SimpleImplementationSLF4J | !XML
+	!AuthenticatedUser | Authentication
+	!ErrorLog | Error
+	!HistoryLog | Trace
