+++ models/uvl/aafms_framework-namesAdapted.uvl
@@ -1,0 +1,100 @@
+namespace AAFMFramework
+
+features
+	AAFMFramework {abstract}	
+		mandatory
+			Metamodels	
+				optional
+					FeatureModel
+					CNFModel
+					DecisionModel
+			System	
+				alternative
+					Linux
+					Win
+
+		optional
+			Packages	
+				mandatory
+					DepMng	
+						mandatory
+							pip	
+								optional
+									pipdeptree
+							setuptools
+
+						optional
+							wheel
+							pkgresources
+
+				optional
+					APISolvers	
+						or
+							pythonsat
+							pycosat
+							pylgl
+							pyPicosat
+							pyglucose
+							pycryptosat
+							satyrn
+							satispy
+							umo
+							pydepqbf
+					ParserTechnology	
+						or
+							antlr4python3runtime
+							SimpleParse
+							larkparser	
+								optional
+									ebnf
+					Utils	
+						or
+							six
+							pybind11
+			AutomatedReasoning	
+				mandatory
+					Solvers	
+						or
+							MiniSAT
+							PicoSAT
+							Glucose
+							CaDiCaL
+							Minicard
+							Maple
+							CryptoMiniSat
+							Lingeling
+							CBC
+							DepQBF
+
+				optional
+					Operations	
+						or
+							ValidModel
+							ValidConfiguration
+							ValidProduct
+							AllProducts
+							DeadFeatures
+							CoreFeatures
+							ErrorDetection
+							ErrorDiagnosis
+			Interoperability	
+				or
+					FeatureIDE
+					SPLOT
+					UVL
+
+constraints
+	MiniSAT => satispy | umo | pythonsat
+	Lingeling => satispy | pylgl | pythonsat
+	CBC => umo
+	PicoSAT => pycosat | pyPicosat
+	Glucose => pythonsat | pyglucose
+	DepQBF => pydepqbf
+	CaDiCaL | Maple | Minicard => pythonsat
+	CryptoMiniSat => pycryptosat
+	satyrn | pyglucose => pybind11
+	pythonsat => six
+	Win => !umo & !pyPicosat
+	CNFModel <=> AutomatedReasoning
+	Linux & pip => pkgresources
+	!UVL | ParserTechnology
