+++ models/uvl/wget.uvl
@@ -1,0 +1,24 @@
+namespace wget
+
+features
+	wget	
+		mandatory
+			base
+
+		optional
+			output_file
+			server_response
+			no_clobber
+			no_dns_cache
+			verbosity	
+				alternative
+					no_verbose
+					verbose
+					quiet
+			no_directories
+			no_host_directories
+			no_http_keep_alive
+			no_cookies
+			save_headers
+			convert_links
+			strict_comments
