+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 python_files =
-    **/*Test.py
+    **/*Test*.py
